Add colors to log-level prefixes
